+++ launch/tier4_simulator_launch/launch/simulator.launch.xml
@@ -141,7 +141,7 @@
   </group>
 
   <group if="$(eval '&quot;$(var localization_sim_mode)&quot;==&quot;api&quot;')">
-    <include file="$(find-pkg-share pose_initializer)/launch/pose_initializer.launch.xml">
+    <include file="$(find-pkg-share autoware_pose_initializer)/launch/pose_initializer.launch.xml">
       <arg name="user_defined_initial_pose/enable" value="false"/>
       <arg name="user_defined_initial_pose/pose" value="[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0]"/>
       <arg name="ndt_enabled" value="false"/>
@@ -159,7 +159,7 @@
       <push-ros-namespace namespace="localization"/>
       <group>
         <push-ros-namespace namespace="util"/>
-        <include file="$(find-pkg-share pose_initializer)/launch/pose_initializer.launch.xml">
+        <include file="$(find-pkg-share autoware_pose_initializer)/launch/pose_initializer.launch.xml">
           <arg name="user_defined_initial_pose/enable" value="false"/>
           <arg name="user_defined_initial_pose/pose" value="[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0]"/>
           <arg name="ndt_enabled" value="false"/>
