+++ localization/autoware_pose_initializer/README.md
@@ -1,8 +1,8 @@
-# pose_initializer
+# autoware_pose_initializer
 
 ## Purpose
 
-The `pose_initializer` is the package to send an initial pose to `ekf_localizer`.
+The `autoware_pose_initializer` is the package to send an initial pose to `ekf_localizer`.
 It receives roughly estimated initial pose from GNSS/user.
 Passing the pose to `ndt_scan_matcher`, and it gets a calculated ego pose from `ndt_scan_matcher` via service.
 Finally, it publishes the initial pose to `ekf_localizer`.
@@ -13,7 +13,7 @@ This node depends on the map height fitter library.
 
 ### Parameters
 
-{{ json_to_markdown("localization/pose_initializer/schema/pose_initializer.schema.json") }}
+{{ json_to_markdown("localization/autoware_pose_initializer/schema/pose_initializer.schema.json") }}
 
 ### Services
 
@@ -51,7 +51,7 @@ If the score of initial pose estimation result is lower than score threshold, ER
 
 ## Connection with Default AD API
 
-This `pose_initializer` is used via default AD API. For detailed description of the API description, please refer to [the description of `autoware_default_adapi`](https://github.com/autowarefoundation/autoware.universe/blob/main/system/autoware_default_adapi/document/localization.md).
+This `autoware_pose_initializer` is used via default AD API. For detailed description of the API description, please refer to [the description of `autoware_default_adapi`](https://github.com/autowarefoundation/autoware.universe/blob/main/system/autoware_default_adapi/document/localization.md).
 
 <img src="../../system/autoware_default_adapi/document/images/localization.drawio.svg" alt="drawing" width="800"/>
