+++ localization/autoware_pose_initializer/media/diagnostic_pose_reliability.png
