+++ localization/autoware_pose_initializer/schema/pose_initializer.schema.json
