+++ localization/autoware_pose_initializer/test/test_copy_vector_to_array.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../src/pose_initializer/copy_vector_to_array.hpp"
+#include "../src/copy_vector_to_array.hpp"
 
 #include <gmock/gmock.h>
 
@@ -20,7 +20,7 @@ TEST(CopyVectorToArray, CopyAllElements)
 {
   const std::vector<int> vector{0, 1, 2, 3, 4};
   std::array<int, 5> array{};
-  copy_vector_to_array<int, 5>(vector, array);
+  autoware::pose_initializer::copy_vector_to_array<int, 5>(vector, array);
   EXPECT_THAT(array, testing::ElementsAre(0, 1, 2, 3, 4));
 }
 
@@ -29,7 +29,7 @@ TEST(CopyVectorToArray, CopyZeroElements)
   const std::vector<int> vector{};
   // just confirm that this works
   std::array<int, 0> array{};
-  copy_vector_to_array<int, 0>(vector, array);
+  autoware::pose_initializer::copy_vector_to_array<int, 0>(vector, array);
 }
 
 TEST(CopyVectorToArray, ThrowsInvalidArgumentIfMoreElementsExpected)
@@ -37,7 +37,7 @@ TEST(CopyVectorToArray, ThrowsInvalidArgumentIfMoreElementsExpected)
   auto f = [] {
     const std::vector<int> vector{0, 1, 2, 3, 4};
     std::array<int, 6> array{};
-    copy_vector_to_array<int, 6>(vector, array);
+    autoware::pose_initializer::copy_vector_to_array<int, 6>(vector, array);
   };
 
   EXPECT_THROW(
